There should be support for the WebWorks io.FileTransfer API.
This implements an acceptable level of support for the upload and
download methods for emulating their behaviour.

Note: However, since this API is supposed to work with cross
origin requests, it makes use of Ripple's cross origin proxy (unless you
run the browser with disabled web security), which needs an overhaul,
itself, and may not work 100% with the upload/download methods. This is
a TODO.